refactor: GPIO pin mask is now a typedef uint16. Passes on STM32.
This is due to the disparity between the dev PC and STM32 compilers. The STM32 compiler enums happened to be 16 bits, dev PC 32 bits. now the
type size is equal on both devices. The test that used to check this has been removed and another added, see diff.

No longer checking size of mask as it is now protected by the size of
uint16.