refactor: move across tests for the operation of write/flush sd
they go in their own module and own functions.

this is one of the steps of separating the
operations in the SerialSnooper_Run into their own module
and individual functions.

the tests were made very complicated simply because of
the fact that the Run() function was doing too many things at once.

This will make it easier to test, modify and maintain the code.

Currently we just split up the operations and their tests.

The next step is to decouple the data that they use which will
reduce complexity of tests further.